# speed up dropout