--- a/scala/dllib/src/main/scala/com/intel/analytics/bigdl/dllib/nn/Dropout.scala
+++ b/scala/dllib/src/main/scala/com/intel/analytics/bigdl/dllib/nn/Dropout.scala
@@ -74,61 +74,15 @@ class Dropout[T: ClassTag](
     }
     if (train) {
       noise.resizeAs(input)
-      if (input.isContiguous()) {
-        if (isResampling) {
-          val noiseData = noise.storage().array()
-          var taskSize = noise.nElement() / Engine.model.getPoolSize
-          var extraTask = noise.nElement() % Engine.model.getPoolSize
-          var allocated = 0
-          val offset = this.output.storageOffset() - 1
-          val data = this.output.storage.array()
-          var i = 0
-          while (allocated < noise.nElement()) {
-            val start = allocated
-            allocated += taskSize
-            if (extraTask > 0) {
-              allocated += 1
-              extraTask -= 1
-            }
-            val end = allocated
-            results(i) = Engine.model.invoke(() => {
-              var k = start
-              while (k < end) {
-                noiseData(k) = if (RNG.bernoulli(1 - p)) {
-                  if (scale) {
-                    data(offset + k) = ev.divide(data(offset + k), ev.fromType[Double](1 - p))
-                    ev.fromType[Double](1.0 / (1 - p))
-                  } else {
-                    ev.fromType[Int](1)
-                  }
-                } else {
-                  data(offset + k) = ev.fromType[Int](0)
-                  ev.fromType[Int](0)
-                }
+      if (isResampling) {
+        noise.bernoulli(1 - p)
 
-                k += 1
-              }
-            })
-            i += 1
-
-          }
-
-          Engine.model.sync(results)
-        } else {
-          this.output.cmul(noise)
+        if (scale) {
+          noise.div(ev.fromType[Double](1 - p))
         }
-        this.output
-      } else {
-        if (isResampling) {
-          noise.bernoulli(1 - p)
-
-          if (scale) {
-            noise.div(ev.fromType[Double](1 - p))
-          }
-        }
-
-        this.output.cmul(noise)
       }
+
+      this.output.cmul(noise)
     } else if (!scale) {
       this.output.mul(ev.fromType[Double](1 - p))
     } else {
@@ -147,39 +101,7 @@ class Dropout[T: ClassTag](
         this.gradInput.resizeAs(gradOutput).copy(gradOutput)
       }
 
-      if (gradInput.isContiguous()) {
-        val noiseData = noise.storage().array()
-        var taskSize = noise.nElement() / Engine.model.getPoolSize
-        var extraTask = noise.nElement() % Engine.model.getPoolSize
-        val gradInputData = gradInput.storage().array()
-        val gradInputOffset = gradInput.storageOffset() - 1
-        var allocated = 0
-        var i = 0
-        while (allocated < noise.nElement()) {
-          val start = allocated
-          allocated += taskSize
-          if (extraTask > 0) {
-            allocated += 1
-            extraTask -= 1
-          }
-          val end = allocated
-          results(i) = Engine.model.invoke(() => {
-            var k = start
-            while (k < end) {
-              gradInputData(gradInputOffset + k) =
-                ev.times(gradInputData(gradInputOffset + k), noiseData(k))
-              k += 1
-            }
-          })
-          i += 1
-        }
-
-        Engine.model.sync(results)
-
-        this.gradInput
-      } else {
-        this.gradInput.cmul(noise)
-      }
+      this.gradInput.cmul(noise)
     } else {
       Log4Error.invalidOperationError(false, "Dropout: backprop only defined while training")
     }