--- a/scala/dllib/src/test/scala/com/intel/analytics/bigdl/dllib/integration/torch/DropoutSpec.scala
+++ b/scala/dllib/src/test/scala/com/intel/analytics/bigdl/dllib/integration/torch/DropoutSpec.scala
@@ -84,4 +84,15 @@ class DropoutSpec extends TorchSpec {
 
     println("Test case : Dropout, Torch : " + luaTime + " s, Scala : " + scalaTime / 1e9 + " s")
   }
-}
+
+  "Dropout performance" should "work" in {
+    val module = new Dropout[Float](0.7, true, false)
+    val input = Tensor[Float](1, 12, 512, 512)
+    val seed = 100
+    input.rand()
+    val start = System.nanoTime()
+    val output = module.forward(input)
+//    val d = input.bernoulli(0.7)
+    println((System.nanoTime() - start) / 1e9)
+  }
+  }