The forward time of dropout (input size (1, 12, 512, 512)) reduce from 0.15s to 0.06s. 
1. use ThreadLocalRandom instead of RandomGenerator. (- 0.03s)
2. move ev.fromType when generating bernoulli data. (- 0.06s)

TODO: 
1. Maybe break the compare test VS torch test.
2. Run the training of ImageNet.